+++ src/main/java/qble2/poe/item/ItemDto.java
@@ -27,10 +27,10 @@ public class ItemDto {
   private String baseType;
 
   @JsonProperty("frameType")
-  private long frameType;
+  private int frameType;
 
   @JsonProperty("ilvl")
-  private long ilvl;
+  private int ilvl;
 
   @JsonProperty("inventoryId")
   private String inventoryId;
@@ -51,16 +51,16 @@ public class ItemDto {
   private Boolean fractured;
 
   @JsonProperty("w")
-  private long w;
+  private int w;
 
   @JsonProperty("h")
-  private long h;
+  private int h;
 
   @JsonProperty("x")
-  private long x;
+  private int x;
 
   @JsonProperty("y")
-  private long y;
+  private int y;
 
   @JsonProperty("descrText")
   private String descrText;
